(#17871) fix nagios types uniqness under ruby 1.9
In ruby 1.8 and earlier, splatting a 1-element array would return the first
element, post 1.9 you always get an array back.

    $ ruby-1.8.7-p371 -e 'a = ["a"]; b = *a; puts b.inspect'
    "a"

    $ ruby-1.9.3-p429 -e 'a = ["a"]; b = *a; puts b.inspect'
    ["a"]

This lead to an issue in the nagios providers where the parameters['name']
would be set to an Array, so simple 'does this resource already exist
checks' would fail as  `parameters['name'] == some_string` would always be false.

This patch changes the assignment from a splatted array to an explicit call to
args.first, which is portable between 1.8 and 1.9 and restores the intended
behaviour.