(#6808) Warn about file content sync no-ops
Normally, a declaration of both `target` and `content` for a File resource in a
manifest will cause a compile error.  However, if a File resource is ensured to
be `present`, it is possible to specify `content` for a path that turns out to
be a symlink on the agent.  Currently, since this situation is not discoverable
untill the agent is applying the catalog, the resource will silently fail to
apply `content` changes.

This patch adds a warning indicating that no sync of `content` will occur.