Docs: Improve documentation of "duration" settings
A user pointed out that "as a duration" was not a helpful way to describe the
formats used by duration settings; although the formats are explained at the top
of the config reference, scrolling shouldn't be necessary, and most readers
wouldn't think to check the top of the file.

This commit puts a small list of the formats in each duration setting, as well
as marking the activerecord storeconfigs settings as deprecated.