Maint: Don't execute facter during spec tests
Previously, the test was trying to test that the 'facts' application
could be invoked and return exit code 0 with meaningful output. However,
it causes the 'facter' terminus to be invoked, which actually runs
facter.

Since we don't care what facts are returned, this commit adds an
expection to make sure the indirection#save method is called and returns
successfully.

This was discovered on win2003 due to local facts being present, which
when evaluated, wrote errors into the log.