(maint) Add an integration spec for agent logging.
Checks which loggers are initialized, and what the log level is for all
combinations of --onetime, --daemonize, --no-daemonize, --verbose,
--debug, and --logdest called on the agent.

A simple patch for #15187 had unintended side effects to logging
behavior which were hard to track down.  This test is intended to give
us visibility on how we expect agent logging to behave and whether or
not this has changed.

There is also a minor change to Puppet::Application#setup_logs that
separates assigning a console log for verbose/debug from setting the
log_level, because this was where one of the bugs was hidden.