(#20768) Refactor base profile logic
This commit refactors the logic for figuring out the base profile
directory, e.g. C:\Users or C:\Documents and Settings, and updates the
acceptance test to use it. The `--helper` option must be used in the
systest harness so that the LOAD_PATH is setup correctly to find the
windows_utils.rb helper, but this should already be the case for
temp_file_utils.rb, etc.