(#20768) Windows can now create the home directory (again)
The previously pending_test method was used, because puppet couldn't
create the home directory when creating the user in ruby 1.9 on windows.
Now that that is resolved, we can remove the windows-specific
pending_test. Note, that the test will still be pending later, since
neither the windows nor useradd providers correctly move the user's home
directory.