(#20967) Fully manage ownership of rundir
When the change from 01777 to 0755 for rundir is applied, the puppet
master can get itself into a situation where it cannot properly write to
its rundir. This occurs when it creats the rundir as root and then later
tries to write the pidfile as the system (usually puppet) user. This is
solved by also making puppet try to manage the ownership of the rundir,
which also makes it consistent with what packages were creating and the
other directories that it tries to manage for the master.