(Maint) Remove unused and unreachable logging configuration
The --centrallogging tried to setup some sort of central aggregation of
logs (see issue #51). However at some point in the past the
--centrallogging flag was disconnected from the internal handling of
this (the code assumed it was called :centrallogs). However, even fixing
that disconnect would not have made this work because the logging system
does not appear to support this kind of setup. The :host logging
destination depends on a class called
Puppet::Network::Client::LogClient, which does not exist.

This simply removes all of that code, including the handling of the
command line options. This results in any user who had been passing
those flags now recieving an error, however, since these were never
documented anywhere that I can find, this is an acceptable risk.