(#21055) Push setup of agent closer to where it is used
This pushes the initialization of different parts of the agent under
various circumstances closer to where the pieces are used. By reducing
the "live-time" of the code the risk of bugs related to partial
initialization is reduced.

Because of the internal refactorings and the tight coupling of the tests
to these internal control flows, many of the tests have been reworked to
take a much more outside-in approach and now mostly (but not completely)
test through a more "public" interface.