(#4689) Set correct blockdevice for nfs mounts
Solaris vfstab has an extra field "blockdevice". The Puppet
Documentation claims that the blockdevice should default to the correct
value and you do not have to set it explicitly. While this may be true
for regular devices the user has to set blockdevice explicitly if the
mount resource describes an NFS share.

The correct entry for NFS shares in /etc/vfstab is to set blockdevice
to "-"