(#6793) Remove trailing slashes from mount resource
The mount system command always displays mount points without trailing
slashes so if a user defines a mount resource with a trailing slash,
puppet will always see the mount resource as unmounted and as a result
will keep remounting the device on every puppet run.

Munge the name and strip away any trailing slashes but make sure that we
do not remove the slash if the mount describes the root directory `/`.
This way the munged name will always equal the mount output and puppet
is able to find it.