(#21168) Set default pass property to "-" on Solaris
The current default value of "0" for the pass property has a different
meaning on Solaris than on other systems. A value of 0 actually depends
on the filesystem type: It will cause the system to not fsck the device
if it is an UFS filesystems and will fsck the device if it is not an UFS
filesystem

(http://docs.oracle.com/cd/E19455-01/805-6331/fsadm-59727/index.html)

I never actually saw a value of "0" in any /etc/vfstab and it seems to
be common to use "-" instead. This will always deactivate filesystem
checking and is also used in many examples (see link above).

Change the default value of the pass property to "-" on Solaris and keep
the default value of zero on other systems.