(maint) Update developer readme for bundler setup.
The README_DEVELOPER.md instructions had some out of date notes
regarding gemspec and initial setup.  While setting up a puppet checkout
with the structure of a locally installed gem is possible, it is more
complicated and requires generation of a puppet.gemspec since this is no
longer included in the repository.

Given that a bundle install/bundle exec puppet cycle now works fine,
this commit updates the readme to provide simple instructions for a
bundler checkout and setup, and includes the steps for generating a
gemspec if a local gem style setup is preferred.