(#21478) Add plymouth-ready to excludes list
An update to Ubuntu 12.04's plymouth package, version 0.8.2-2ubuntu31,
introduces a new upstart service, plymouth-ready in order to fix
the bug detailed here:
https://bugs.launchpad.net/ubuntu/+source/lightdm/+bug/982889

This new service uses instances, which puppet does not yet fully
support, and attempting to retrieve the status of this job causes
puppet to fail (for example when running "puppet resource service").

Adding plymouth-ready to the excludes list allows puppet to output
the status of services that it can handle competently while avoiding
the service introduced by this package.