(Maint) Test that the query string is parsable
Rather than relying on all sorts of string checks, it is better to check
that a parameter parsing library is able to parse the generated string
back into the expected data. This changes the tests to use CGI::parse to
parse the query_string and then can check that the resulting hash has
the data that is expected.