(#18295) add title to start call in windows exe package provider
When called without a title parameter, the windows 'start' function
collects any additional command line switches to itself rather than
passing them along to the command to be started.

The existing exe_package.rb, in it's initial and only commit
07931e63c7d76fcec331946921b304438c38b469, had the "Title" parameter set
for uninstalls. It did not have the "Title" parameter set for installs.
Since uninstalls almost always require parameters, but install do not,
I assume this was an unintentional omission.

This resolves the issue by adding a title bar description with the
string 'puppet-install' to windows executed for installing .exe
packages.