(#16856) Add validator, checker, issue markers, and tests for the binding model This commit adds the initial implementation of the following classes:
BindingsChecker
Checks the validity of a bindings model

BindingIssues
Contains all declaration of all issues that the BindersChecker can
report

BindingsLabelProvider
String representation of all binder model elements

BindingsValidatorFactory
The factory for creating the checker and associate it with the
correct support classes.

Unit tests are provided for the checker and factory. They do in turn
use the label provider and issue.