(#16856) Improve API and documentation of Injector.
This adds the ability to optionally give a block to
or just result. The lookup returns what the block returns if block is
present.

This is good because it make it easier to handle default values when
a value is missing, or to conveniently invoke a custom producer method
without requiring a temporary variable.

Includes tests.