PSON serialization for `Resource::Status` missing time field
Prior to this commit, the PSON serialization for the class
`Puppet::Resource::Status` was missing the `time` field.  This
would result in `nil` values in reports, which was causing
failures in the PuppetDB report processor.  This commit
simply adds in the field to the serialization, and adds
corresponding test coverage.