(#18255) - Follow http redirections
This patch adds support for HTTP redirection in the Puppet HTTP
client.

The http client includes a redirection limit, and recreate the
redirected connection with the same certificates and store as the
original (as long as the redirect new location is ssl protected).

This patch allows redirection following for all HTTP methods, which
in the Puppet ecosystem shouldn't be an issue.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>

Use a loop - temp commit