(#11317) - Do not ever skip resources appliable to both host and device
Since the puppet device feature introduction some types are labelled
as usable on host only, device only or both.

The schedule type is the only type that was labelled 'both'. Unfortunately
the logic in the Transaction#skip? was flawed and this resulted in all
the 'both' kind of resources to always being skipped, which essentially
meant that the Schedule type was non-working for a long time.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>