(#16856) Make instantiating producer work like assisted inject
Assisted inject and instance producer now works the same way; if there
is a class method :inject this has higher precedence when creating the
instance. Assisted inject now also passes arguments to the #inject
method.

Assisted inject refactored to handle this.

It was not logical that instance production did not use the #inject
class method if one was present.