(#19727) Collapse multiple leading slashes
Puppet 2.6.5 to 2.7.19 collapsed multiple leading slashes. This behavior
was changed in 2.7.20 in commit 912ed346 as a result of no longer
calling `File.split`.

This commit reinstates the behavior that existed previously and adds
tests for both double and triple leading slashes.

Paired-with: Rob Reynolds <rob@puppetlabs.com>
Paired-with: Kylo Ginsberg <kylo@puppetlabs.com>