Clean up based on Justins review
The core of this branch has focused on the install tests. However,
early commits to expand the suite introduced inconsistencies in
other components of the suite.

This commit specifically addresses these inconsistencies in order
to bring these files into parity with the tests in the install
directory.

Thanks to @justinstoller for bringing these to my attention via
his comments on July 9.