(#6112) Acceptance test handles host with uninitialized ssl
Running the acceptance test on Jenkins discovered a third case, and
helped use clarify behavior for the expected cases on an initialized
master/agent.  The third case is a host which has not initialized yet
(has no ssl info created or cached) which is similar to what happens
when a master initializes through a 'puppet master' call, but this is
through a 'puppet generate' call and what the node should be capable of
as a master or agent afterwords is ambiguous.

Acceptance test now checks this case, showing that a cert does generate
in this instance, and leaves the host in an ambiguous state of ssl
configuration where they cannot authenticate with the master afterwords.